+++ nfm-world/mad/gameplay/BasePhase.cs
@@ -19,11 +19,25 @@ public abstract class BasePhase
         MouseDownThisFrame = false;
     }
 
+    /// <summary>
+    /// Use <see cref="G"/> here to draw 2D overlays.
+    /// Use <see cref="Scene"/> here to draw 3D content.
+    /// </summary>
     public virtual void Render()
     {
     }
 
+    /// <summary>
+    /// Use ImGui methods in here.
+    /// </summary>
     public virtual void RenderImgui()
+    {
+    }
+
+    /// <summary>
+    /// Renders after 2D overlays. Use to draw 3D content over 2D content.
+    /// </summary>
+    public virtual void Render3DOverlays()
     {
     }
 
@@ -62,10 +76,6 @@ public abstract class BasePhase
     }
 
     public virtual void WindowSizeChanged(int width, int height)
-    {
-    }
-
-    public virtual void RenderAfterSkia()
     {
     }
 }
