+++ nfm-world/mad/gameplay/BaseRacePhase.cs
@@ -194,10 +194,10 @@ public abstract class BaseRacePhase(GraphicsDevice _graphicsDevice) : BaseStageR
         if(DebugDisplay) {
             RenderMessages();
             G.SetColor(new Color(0, 0, 0));
-            G.DrawString($"Render: {Program._lastFrameTime}ms", 100, 100);
-            G.DrawString($"Tick: {Program._lastTickTime}μs", 100, 120);
+            G.DrawString($"Render: {WorldGame._lastFrameTime}ms", 100, 100);
+            G.DrawString($"Tick: {WorldGame._lastTickTime}μs", 100, 120);
             G.DrawString($"Power: {CarsInRace[0]?.Mad?.Power:0.00}", 100, 140);
-            G.DrawString($"Ticks executed last frame: {Program._lastTickCount}", 100, 160);
+            G.DrawString($"Ticks executed last frame: {WorldGame._lastTickCount}", 100, 160);
         }
     }
