+++ nfm-world/mad/renderer/RenderState.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using MoonWorks.Graphics;
 using GpuBuffer = MoonWorks.Graphics.Buffer;
 
@@ -11,17 +12,59 @@ namespace nfm_world;
 public static class RenderState
 {
     /// <summary>The command buffer for the current frame.</summary>
-    public static CommandBuffer Cmd;
+    [AllowNull]
+    public static CommandBuffer Cmd
+    {
+        get => field ?? ThrowAccessedOutsideFrame<CommandBuffer>(nameof(Cmd));
+        set;
+    }
 
     /// <summary>The active render pass (changes between shadow/main passes).</summary>
-    public static RenderPass Pass;
+    [AllowNull]
+    public static RenderPass Pass
+    {
+        get => field ?? ThrowAccessedOutsideRenderPass<RenderPass>(nameof(Pass));
+        set;
+    }
 
     /// <summary>The swapchain backbuffer texture for the current frame.</summary>
-    public static Texture Backbuffer;
+    [AllowNull]
+    public static Texture Backbuffer
+    {
+        get => field ?? ThrowAccessedOutsideFrame<Texture>(nameof(Backbuffer));
+        set;
+    }
 
     /// <summary>The main depth texture for depth testing.</summary>
-    public static Texture MainDepthTexture;
+    [AllowNull]
+    public static Texture MainDepthTexture
+    {
+        get => field ?? ThrowAccessedOutsideFrame<Texture>(nameof(MainDepthTexture));
+        set;
+    }
 
-    /// <summary>Set to true when Scene has rendered 3D content to the backbuffer this frame.</summary>
-    public static bool SceneRenderedThisFrame;
+    public static void BeginDraw(CommandBuffer renderCmd, Texture backbuffer, Texture mainDepthTexture)
+    {
+        Cmd = renderCmd;
+        Backbuffer = backbuffer;
+        MainDepthTexture = mainDepthTexture;
+    }
+
+    public static void EndDraw()
+    {
+        Cmd = null;
+        Pass = null;
+        Backbuffer = null;
+        MainDepthTexture = null;
+    }
+    
+    private static T ThrowAccessedOutsideFrame<T>(string paramName)
+    {
+        throw new InvalidOperationException($"Accessed {nameof(RenderState)}.{paramName} outside of a frame, must call {nameof(BeginDraw)} first!");
+    }
+
+    private static T ThrowAccessedOutsideRenderPass<T>(string paramName)
+    {
+        throw new InvalidOperationException($"Accessed {nameof(RenderState)}.{paramName} outside of a render pass!");
+    }
 }
