+++ nfm-world/mad/ui/ModelEditor.cs
@@ -2355,9 +2355,9 @@ public class ModelEditorPhase : BasePhase
         GameSparker.ExitEditor();
     }
     
-    public override void RenderAfterSkia()
+    public override void Render3DOverlays()
     {
-        base.RenderAfterSkia();
+        base.Render3DOverlays();
         
         var tab = ActiveTab;
         if (!_isOpen || tab == null || tab.Object == null) return;
